feature/storybookSetup: typography
