+++ src/components/common/Button/button.css
@@ -5,7 +5,6 @@
   border-radius: 3em;
   font-weight: 700;
   line-height: 1;
-  font-family: 'Nunito Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
 }
 .storybook-button--primary {
   background-color: #1ea7fd;
